# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Saina Gallery RESTful API",
        "description": "SainaGallery API documentation"
    },
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/sellerlogin": {
            "post": {
                "tags": [
                    "api"
                ],
                "summary": "Seller Login",
                "description": "Seller Login Route",
                "operationId": "sellerlogin",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "body",
                        "description": "Login Parameters",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Invalid Status Value"
                    }
                }
            }
        },
        "/api/sellerregister": {
            "post": {
                "tags": [
                    "api"
                ],
                "summary": "Register a New Seller",
                "description": "User Can Register using this Route as Seller",
                "operationId": "sellerregister",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "register",
                        "description": "Register a New User",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/seller"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Invalid Status Value"
                    }
                }
            }
        },
        "/api/addsuperadmin": {
            "post": {
                "tags": [
                    "api"
                ],
                "summary": "Add Superadmin",
                "description": "Register new Superadmin",
                "operationId": "registersuperadmin",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "addsuperadmin",
                        "in": "body",
                        "description": "Register new Superadmin",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Invalid Status Value"
                    }
                }
            }
        },
        "/api/province": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get one Province",
                "description": "Get one Province",
                "operationId": "getoneprovinces",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id",
                        "description": "Province id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/api/city": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get one City",
                "description": "Get one City",
                "operationId": "getonecity",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id",
                        "description": "City id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/api/category": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get one Category",
                "description": "Get one Category",
                "operationId": "getonecategory",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id",
                        "description": "Category id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/api/gallery": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get one Gallery",
                "description": "Get one Gallery",
                "operationId": "getonegallery",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id",
                        "description": "Gallery id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "definitions": {
        "login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "imanw77@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            },
            "xml": {
                "name": "login"
            }
        },
        "users": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string",
                    "example": "Iman - [REGISTER]"
                },
                "lastName": {
                    "type": "string",
                    "example": "Veisi - [REGISTER]"
                },
                "password": {
                    "type": "string",
                    "example": "123456 - [ONLY ON REGISTER]"
                },
                "email": {
                    "type": "string",
                    "example": "imanw77@gmail.com - [ONLY ON REGISTER]"
                }
            },
            "xml": {
                "name": "users"
            }
        },
        "seller": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string",
                    "example": "Iman"
                },
                "lastName": {
                    "type": "string",
                    "example": "Veisi"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "email": {
                    "type": "string",
                    "example": "imanw77@gmail.com"
                }
            },
            "xml": {
                "name": "seller"
            }
        },
...
    }
}